# docs/admin/swagger.yaml
definitions:
...
  request.CompatibilityRule:
    properties:
      condition:
        type: string
      errorMessage:
        type: string
      isActive:
        type: boolean
      sourceComponentTypeCode:
        type: string
      sourceKey:
        type: string
      targetComponentTypeCode:
        type: string
      targetKey:
        type: string
      valueType:
        type: string
    type: object
  request.ComponentType:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  response.AuditTrail:
    properties:
      id:
        type: integer
      userName:
        type: string
    type: object
  response.CompatibilityRule:
    properties:
      condition:
        type: string
      createdAt:
        type: string
      createdById:
        $ref: '#/definitions/response.AuditTrail'
      errorMessage:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      modifiedAt:
        type: string
      modifiedById:
        $ref: '#/definitions/response.AuditTrail'
      sourceComponentTypeCode:
        type: string
      sourceKey:
        type: string
      targetComponentTypeCode:
        type: string
      targetKey:
        type: string
      valueType:
        type: string
    type: object
  response.ComponentType:
    properties:
      code:
        type: string
      createdAt:
        type: string
      createdById:
        $ref: '#/definitions/response.AuditTrail'
      description:
        type: string
      id:
        type: integer
      modifiedAt:
        type: string
      modifiedById:
        $ref: '#/definitions/response.AuditTrail'
      name:
        type: string
    type: object
...
  response.WebResponse:
    properties:
      code:
        type: integer
      content: {}
      message:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
info:
  contact: {}
  description: Swagger documentation for admin
  title: Admin API
  version: "1.0"
paths:
  /api/admin/compatibility-rules:
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CompatibilityRule'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Search compatibility rules
      tags:
      - Compatibility Rules
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CompatibilityRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Create compatibility rules
      tags:
      - Compatibility Rules
  /api/admin/compatibility-rules/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete compatibility rules
      tags:
      - Compatibility Rules
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CompatibilityRule'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Get compatibility rules by id
      tags:
      - Compatibility Rules
    put:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CompatibilityRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Update compatibility rules
      tags:
      - Compatibility Rules
  /api/admin/component-types:
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ComponentType'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Search component type
      tags:
      - Component Type
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ComponentType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Component ype
      tags:
      - Component Type
  /api/admin/component-types/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete component type
      tags:
      - Component Type
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ComponentType'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Get component type by id
      tags:
      - Component Type
    put:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: path
        name: id
        required: true
        type: integer
      - description: ' '
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ComponentType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.WebResponse'
      security:
      - ApiKeyAuth: []
      summary: Update component type
      tags:
      - Component Type
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WebResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.WebResponse'
      summary: Login
      tags:
      - Auth
...
swagger: "2.0"